turn writer in module
